Update `frame->ip` before calling runtime_error() in run()
Since ip is in a separate (register) variable, runtime_error() does not have access to the latest value of ip.
This often causes an error to be rported on 'line -2', since -2 is returned by get_line() if the provided index
is outside the range of indices stored in the line list.